Invalid Create Table Fixed

+++ src/main/resources/sql/db/create-table-games.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS games (
-    id BIGINT PRIMARY KEY AUTOINCREMENT,
+    id INTEGER PRIMARY KEY AUTOINCREMENT,
     name TEXT not null,
     steam_id TEXT,
     target_price numeric (6, 2), -- MAX: 9999,99
